Successfully entered user-mode
Unfortunately, we commit a variety of shameful hacks along the way, all of
which we will doubtless have to answer for some day.  Among our varied
offenses... we do not setting EFLAGS correctly (we use the current value in
the kernel's EFLAGS), we do not zero the various GP registers, do not store
the kernel stack in ESP0, and we DID have to mark the page directory
entries for ALL user page tables as user-accessible (this might be ok; more
testing is needed).  There might be (read: likely are) other misdeeds that
I cannot recall.