Fixed loader; cleaned up mem map stuff
The loader allocates both RX sections together, then both RW sections.  For
simplicity, we assume that the .text section is lower in memory than
.rodata, and that .data is lower in memory than .bss.  Furthermore, the vm
allocator takes a vm_info_s pointer instead of a pointer to the page
directory.  Unfortunately, getting this to work required lots of small
changes all over the place.  It all seems to work, at least for schizo and
introspective, but we've got lots of clean up to do...