Fixed N_ARY_SYSCALL
Fixed the two bugs that were causing us to fail exec_nonexist:

1) Exec zeroed the vm of the current process and then loaded the file, but if
the file does not exist and then you try to continue executing the current
program, you get a page fault.  I solved this by adding a file validation at the
beginning of exec before zeroing out memory. However, the function that I wrote,
validate_file() is redudant with regards to the other loader functions so we
should probably parameterize it to take an elf struct because then it can be
called by the other functions. I was reluctant to do this because I didn't want
to do any elf related operations in syscalls, but I suppose with exec you are
loading an elf...

2) The asm wrapper for exec N_ARY_SYCALL had a lot of bugs
    a. It saved and restored the caller-saved register ECX but not the callee
    saved register EDI.
    b. Data segments were being pushed/popped NOT in symmetric order.  It
    doesn't matter since they all contain the same values, but we look like
    amateurs doing it that way.
    c. Rep takes the STARTING address for storage so we need store ESP - 8 in
    EDI, not ESP.  Because of this oversight we were clobbering the values
    restored to the data segment registers, EDI and EBP, and the iret wasn't
    working properly because the ESP was 8 bytes higher than it should have
    been.

Also, although I did not test it, UNARY_SYSCALL had a small bug: it was pushing
an argument for a call and then not realigning ESP upon return.

I forgot to document this in my previous commits, but I ran some of the basic
tests (see config.mk to see the specific names) and debugged our code so we are
passing all of them but coolness which tests ctx switching, exec and fork.  I
need to look at the exec code and how it interacts with my naive thread list
since I coded up the thread list with fork in  mind.  Being able to run exec and
fork together will be a good benchmark.