Successfully modularized vm_alloc()
ZFOD now works on basic exec and fork tests.  We probably want to look back at
how we load bss/data because the code for handling bss and data overlapping a
page is ugly.  Also, I'm a bit concerned about how we are handling holes between
regions in the sense that because we are mapping them in and zeroing them out
the user will not page fault if he/she accesses those regions.  Perhaps we can
provide more fine grained permissions by generalizing my zfod function to take
an attrs parameter so that the memory is read only but NOT writeable.