Added live children member to task
Vanish now sets the parent of its live children to be init, and gives its
unreaped dead children to init.  In order to test this, we need to have a fixed
loader so we can actually run init.  Also, there's a LOT of locking in vanish,
and it's probably something we want to rethink.