Fixed bug in usr stack init
I ported the frame allocator code over and it worked for basic execing and
forking tests but not for shell.  That's because shell- unlike our small tests-
uses more than one stack frame.  This was problematic because usr_stack_init
calls set vm_attrs, which sets the page table attributes for the entire region.
And since the stack region is an entire tome with every page table entry mapped
to the zfod page, the zfod page was now writeable from the perspective of the
root shell thread.  When the shell forked and called copy_page, copy_page
doesn't copy zfod pages and instead copies the page table entry.  After fork the
parent would continue executing and call wait and then deschedule itself, and
the top of the user stack had the return address of wait.  Thus when the child
return from fork, since it was sharing the zfod stack frame with its parent, it
would return to the return address pushed on by wait rather than fork.

This was challenging to bebug because when I first compared the stacks for the
child and the parent, they were each backed by their own respective physical
frame because it was the first page of the stack. However, I'm still a bit
confused, because I added a conditional print for when you were copying a page
table entry backed by a zfod page, and it only triggered once.