Reverted the page table functions to macros
Now we use macros (again) instead of bit-fields to implement and manipulate
our page tables.  I've gotten rid of the old bit-field stuff entirely, but
the interface still needs some work.  In particular, we should have macros
for setting a PTE's attributes and address, and we should probably get rid
of the init_pte(...) function, as it's functionality is covered by the
PACK_PTE(...) macro.