Customized fault wrapper macros for swexn
There are now two fault wrapper macros, one for faults when the processor pushes
an error code on the stack, and one for when the processor does not.  It is
neccessary to have two different handlers because we need to adjust esp
accordingly for the iret.  I also got rid of the prologue and epilogue so that
we can cast the stack contents to a ureg_t.