Added the remaining hurdle tests
I added the stability tests to config.mk and will run these over night.  I ran
the basic tests, and we pass all of them except for halt() and yield_desc_mkrun.
I also added and ran the remaining solidity tests, and we pass all of them
except for:

1) We fail an assertion in fork_exit_bomb.
2) We fail the same assertion STARTING fork_bomb.
3) Make_crash- I'm not sure of our status, the test is still running.
4) Register test- This tests saved registers across context switches.  We should
run this over night.