Identified a locking bug
Our implementation of locking makes it impossible to try to acquire two locks
consecutively (which is a good thing).  However, if we are blocking waiting for
a lock, in that code to blocking, we cannot attempt to acquire another lock.
This problem manifested itself when a thread attempted to block on malloc()
because that lock was already held, but when the thread tried to free its node
from the runnable list, it then blocked on the lock for free and tried to remove
itself from the runnable queue even though it was already free.