Added p2 libthread tests
I added them to git because it was a bit annoying to copy all the tests over and
add them to the config.mk, and I wasn't sure if you already had.  Plus, if we
run make update, I think it will overwrite all the tests in 410user, which would
also be a bit irksome.