Pages are now zeroed upon freeing
Pages are zeroed in free_frame() since that function is called by both
free_page() and pg_tbl_free().  However, the implementation is not incredibly
slick- since we store the pointer to the free list in the lowest word, then we
need to zero this upon page allocation in alloc_page_table() and
alloc_page_table_really().  So an alternative to the current implementation
is to instead just zero everything then.