Added wrapper for fault handlers
This wrapper takes the kernel fault handler as an argument and proceeds as
follows:
1) calls the kernel handler, allowing the kernel to attempt to silently handle
the fault
2) calls the swexn (if installed)
3) slaughters the faulting thread.

However, currently I'm trying to push \handler, but this just pushes the first 4
bytes of worth of instructions of the function, rather than the addrss of the
function itself.  How can this be fixed?