Added checks for false negatives
The artificial constraint on memory made me realize we are passing a lot of
tests we should be failing, either because fork() fails or because a faulting
thread is killed by the kernel.  I added breaks in fork() when it fails and
created a list of programs that should be killed and check that list every time
a thread is killed.

I also added a lot of isolation tests that I ran last time when I was attempting
to reason about the memory leak.  These tests essentially run separate 410 user
tests infinitely.  The fact that these tests run over night without crashing,
and make_crash is the only test that crashes, has led me to conclude that our
problem is our zombie threads rather than a memory leak.