Cleaned up thread management system calls.
There was a spec violation in sys_sleep(...) (not returning immediately for
ticks=0), and a genuine bug in sys_deschedule(...)/sys_make_runnable(...)
(not checking that the descheduled thread was descheduled due to a call to
sys_deschedule(...)).  Also, I fixed the cooperative test so it actually
forks threads and not processes--i.e. it should actually work now.